src/requirements-badcorp.txt
cachetools==2.0.1
click==7.0
python-dotenv==0.10.1
